Updated tests and imports
-Moved njit and prange default definition to root module
-Changed tests to handle Numba presence
